:wrench: fix types

+++ src/Trampoline.cpp
@@ -248,7 +248,7 @@ static const size_t kAlignmentSize      = 64;         // 64
             }
         }
         if(!trampoline) {
-            TrampolineInfo info = {NULL, kAllocationSize, ceil((double)alloc_size / kAlignmentSize) * kAlignmentSize};
+            TrampolineInfo info = {NULL, kAllocationSize, (size_t)ceil((double)alloc_size / kAlignmentSize) * kAlignmentSize};
             trampoline = info.base = (unsigned char*)TrampolineAllocate((const unsigned char*)src, info.size);
             g_trampolines.push_back(info);
         }
